[ADDED] Hilt to DI on Android ViewModels Acked-by: Ricardo Taboada <tabo.odricar@gmail.com>
As mentioned on the Documentation in order to have a good application is needed to have a good architecute. Having classes to handle the Firebase Services will allow us to maintain such information outside UI logic.
This allow with DI will allows to not worried to build an create object instance manually and delegate such task to Kotlin